+++ inject/inject-app/src/main/scala/com/twitter/inject/app/internal/ConsoleWriterModule.scala
@@ -1,0 +1,9 @@
+package com.twitter.inject.app.internal
+
+import com.twitter.inject.TwitterModule
+import com.twitter.inject.app.console.ConsoleWriter
+
+object ConsoleWriterModule extends TwitterModule {
+  override def configure(): Unit =
+    bind[ConsoleWriter].toInstance(new ConsoleWriter(Console.out, Console.err))
+}