+++ inject/inject-app/src/test/java/com/twitter/inject/app/EmbeddedAppIntegrationJavaTest.java
@@ -6,6 +6,7 @@ import org.junit.Assert;
 import org.junit.Test;
 
 import com.twitter.app.FlagParseException;
+import com.twitter.inject.app.console.ConsoleWriter;
 
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.mock;
@@ -71,5 +72,33 @@ public class EmbeddedAppIntegrationJavaTest extends Assert {
 
     assertEquals(sampleJavaApp.getSampleServiceResponse(), "hi mock");
     reset(mockSampleJavaAppService);
+  }
+
+  @Test
+  public void testEmbeddedAppWithTestConsoleWriter() {
+    AbstractApp myApp = new AbstractApp() {
+      @Override
+      public void run() {
+        ConsoleWriter console = injector().instance(ConsoleWriter.class);
+        console.out().println("Hello, World!");
+        console.err().println("Oh, No!");
+        System.out.println("This doesn't get captured");
+        System.err.println("This also does not get captured");
+        info(() -> "this will not be captured");
+        error(() -> "this will also not be captured");
+      }
+    };
+
+    TestConsoleWriter testConsole = new TestConsoleWriter();
+    EmbeddedApp app = new EmbeddedApp(myApp);
+    app.bindClass(ConsoleWriter.class, testConsole);
+
+    assertEquals(testConsole.inspectOut(), "");
+    assertEquals(testConsole.inspectErr(), "");
+
+    app.main();
+
+    assertEquals(testConsole.inspectOut(), "Hello, World!\n");
+    assertEquals(testConsole.inspectErr(), "Oh, No!\n");
   }
 }