+++ CHANGELOG.rst
@@ -7,6 +7,11 @@ Note that ``RB_ID=#`` and ``PHAB_ID=#`` correspond to associated message in comm
 Unreleased
 ----------
 
+22.4.0
+------
+
+No Changes
+
 22.3.0
 ------